requirements.txt
python-telegram-bot==22.1
python-dotenv==1.0.0
schedule==1.2.0
openai==1.82.0 